Minimal version with most devices removed
This is a first pass at demonstrating that it is practical to use a stripped-down version of ngspice in the browser, after compilation with emscripten. Most of the bulk of the generated javascript comes from the different device types, most of which are not needed for educational use cases. Removing these devices brings the minified Javascript down to a reasonable size (~2.5 MB).